requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
tqdm>=4.62.0
requests>=2.26.0
python-dotenv>=0.19.0

# Soccer analytics
itscalledsoccer>=0.2.0

# API framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
alembic>=1.7.0
databases>=0.6.0
asyncpg>=0.25.0
aiosqlite>=0.17.0 # For SQLite async support

# Authentication & HTTP
bcrypt>=3.2.0
httpx>=0.24.0

# Additional utilities
beautifulsoup4>=4.10.0  # For scraping
rapidfuzz>=2.0.0  # For fuzzy string matching

# ML utilities
autogluon>=0.8.0  # For Python 3.12 support
scikit-learn>=1.3.0  # Fallback ML
lightgbm>=3.3.0  # Additional ML algorithm

# Development dependencies
pytest>=6.2.0
pytest-asyncio>=0.15.0
black>=21.9b0
flake8>=3.9.0

# Authentication
PyJWT>=2.8.0